Restructure
Took the GUI input out of the business logic, resized the XAML for
chicken and added another panel.  Added events to chickenwindow.xaml.cs
so that the dice can be rolled when clicked.  Moved the constructor for
Random so that the numbers produced were not identical.  Made some class
fields public so that they can be accessed.