changed strings to const, set up scoring
Made the position strings constants under the static class ScoringClass.
Fixed logic error in roll method in ChickenLogic.  Changed the methods
to change the graphics on each roll from a foreach loop iterating
through the buttons.  Added scoring capabilities to the ScoringClass
that add the additional score to a string that is displayed.  Added
extra labels to ChickenWindow so that a current score and a total score
can be displayed.